requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
greenlet==2.0.2
hashids==1.3.1
idna==3.4
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==2.0.4
